rename files and update description
