+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/couple/amount/without_children.yaml
@@ -1,12 +1,13 @@
-description: Quebec limits its adapted work premium credit to eligible couple without child at or below this amount.
+description: Quebec provides the adapted premium credit up to this amount to eligible couples without children.
 values:
   2022-01-01: 3_075.78
 metadata:
   unit: currency-CAD
-  label: Quebec maximum adapted annual work premium amount for couple without children
+  period: year
+  label: Quebec maximum adapted work premium amount for couple without children
   reference:
     - title: Revenu Quebec - Amount of the Work Premium Tax Credit
       href: https://www.revenuquebec.ca/en/citizens/tax-credits/work-premium-tax-credits/tax-credit-amounts/
     - title: Parameters of the Personal Income Tax System for 2022 - Table 3
       href: http://www.finances.gouv.qc.ca/MAJ2021/documents/en/AUTEN_IncomeTax2022.pdf#page=15
-    # Legal reference unavaliable
+    # Legal reference unavailable
