+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/work_income_requirement.yaml
@@ -1,4 +1,4 @@
-description: Quebec limits its adapted work premium credit to one's annual work income above this amount.
+description: Quebec limits its adapted work premium credit to people with work income at least this amount.
 values:
   2022-01-01: 1_200
 metadata:
