+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/general_work_premium/couple/amount/with_children.yaml
@@ -1,12 +1,13 @@
-description: Quebec limits its work premium credit to eligible couple with at least one child at or below this amount.
+description: Quebec provides a work premium credit up to this amount to eligible couples with at least one child.
 values:
-  2022-01-01: 3_449.5
+  2022-01-01: 3_449.50
 metadata:
   unit: currency-CAD
-  label: Quebec maximum anual work premium amount for couple with at least one child
+  period: year
+  label: Quebec maximum work premium amount for couple with at least one child
   reference:
     - title: Revenu Quebec - Amount of the Work Premium Tax Credit
       href: https://www.revenuquebec.ca/en/citizens/tax-credits/work-premium-tax-credits/tax-credit-amounts/
     - title: Parameters of the Personal Income Tax System for 2022 - Table 3
       href: http://www.finances.gouv.qc.ca/MAJ2021/documents/en/AUTEN_IncomeTax2022.pdf#page=15
-    # Legal reference unavaliable
+    # Legal reference unavailable
