+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/general_work_premium/single/amount/person_living_alone.yaml
@@ -1,12 +1,13 @@
-description: Quebec limits its work premium credit to individual living alone at or below this amount.
+description: Quebec provides a work premium credit up to this amount to eligible singles living alone.
 values:
   2022-01-01: 1_025.21
 metadata:
   unit: currency-CAD
-  label: Quebec maximum anual work premium amount for individual living alone
+  period: year
+  label: Quebec maximum work premium amount for individual living alone
   reference:
     - title: Revenu Quebec - Amount of the Work Premium Tax Credit
       href: https://www.revenuquebec.ca/en/citizens/tax-credits/work-premium-tax-credits/tax-credit-amounts/
     - title: Parameters of the Personal Income Tax System for 2022 - Table 3
       href: http://www.finances.gouv.qc.ca/MAJ2021/documents/en/AUTEN_IncomeTax2022.pdf#page=15
-    # Legal reference unavaliable
+    # Legal reference unavailable
