+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/qc_work_premium_credit.yaml
@@ -1,21 +1,8 @@
-- name: Quebec work premium credit for household eligible for adapted work premium credit
+- name: Quebec work premium credit 
   period: 2022
   input:
-    qc_general_work_premium_amount: 200
-    qc_adapted_work_premium_amount: 300
-    qc_work_premium_supplement_amount: 300
-    province_code: QC
-  output:
-    qc_work_premium_credit: 600 # max(200, 300)+300
-
-- name: Quebec work premium credit for married household eligible for adapted work premium credit
-  period: 2022
-  input:
-    qc_general_work_premium_amount: 200
-    qc_adapted_work_premium_amount: 0
-    qc_work_premium_supplement_amount: 300
+    qc_work_premium_credit_base: 200
+    qc_work_premium_credit_supplement: 300
     province_code: QC
   output:
     qc_work_premium_credit: 500 #200+300
-
-
