+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/qc_work_premium_eligible.yaml
@@ -8,7 +8,7 @@
     is_full_time_student: false
     province_code: QC
   output:
-    qc_work_premium_eligibility: true
+    qc_work_premium_eligible: true
 
 - name: Quebec work premium tax credit requirements for people over age limit but is full time student
   period: 2022
@@ -20,7 +20,7 @@
     is_full_time_student: true
     province_code: QC
   output:
-    qc_work_premium_eligibility: false
+    qc_work_premium_eligible: false
 
 - name: Quebec work premium tax credit requirements for people below age limit and unmarried
   period: 2022
@@ -32,7 +32,7 @@
     is_full_time_student: false
     province_code: QC
   output:
-    qc_work_premium_eligibility: false
+    qc_work_premium_eligible: false
 
 - name: Quebec work premium tax credit requirements for people below age limit and married with children and emancipated
   period: 2022
@@ -44,7 +44,7 @@
     is_full_time_student: false
     province_code: QC
   output:
-    qc_work_premium_eligibility: true
+    qc_work_premium_eligible: true
 
 - name: Quebec work premium tax credit requirements for people below age limit and married with children and but not emancipated
   period: 2022
@@ -56,7 +56,7 @@
     is_full_time_student: false
     province_code: QC
   output:
-    qc_work_premium_eligibility: false
+    qc_work_premium_eligible: false
 
 - name: Quebec work premium tax credit requirements for people below age limit and married without children
   period: 2022
@@ -68,4 +68,4 @@
     is_full_time_student: false
     province_code: QC
   output:
-    qc_work_premium_eligibility: false
+    qc_work_premium_eligible: false
