+++ policyengine_canada/variables/gov/provinces/qc/tax/income/credits/work_premium/qc_work_premium_credit.py
@@ -8,20 +8,4 @@ class qc_work_premium_credit(Variable):
     definition_period = YEAR
     defined_for = ProvinceCode.QC
 
-    def formula(household, period, parameters):
-        qc_general_work_premium_amount = household(
-            "qc_general_work_premium_amount", period
-        )
-        qc_adapted_work_premium_amount = household(
-            "qc_adapted_work_premium_amount", period
-        )
-
-        # If you are entitled to both the work premium and the adapted work premium, you will receive the greater of the two
-        credit_amount = max_(
-            qc_general_work_premium_amount,
-            qc_adapted_work_premium_amount,
-        )
-
-        supplement = household("qc_work_premium_supplement_amount", period)
-
-        return credit_amount + supplement
+    adds = ["qc_work_premium_credit_base", "qc_work_premium_credit_supplement"]
